dockerfiles/offline-demo/pyproject.toml
[project]
name = "vllm-offline-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.2.8",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "ray>=2.46.0",
    "typer>=0.15.3",
    "vllm>=0.8.5.post1",
]
